+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New configuration parameter `AllowAdjacentOneLineDefs` for `EmptyLineBetweenDefs`.
 * `RedundantSelf` cop does auto-correction
+* New cop `MultilineBlockChain` keeps track of chained blocks spanning multiple lines.
 
 ### Changes